// src/app/models/user.ts
export class User {
    id: number;
    email: string;
    name: string;
    lastName: string;
    password: string;
    token: string;
    avatar: string;
    sex: string;
    telephone: string;
    verified: boolean;
    role: string;
    validationCode: string;
    startDate : Date;
    status: string;

}